# Alerting: Improve performance of tupleLablesToLabels function